auth

+++ app/Http/Controllers/AuthorController.php
@@ -1,9 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\models\Author;
 use Illuminate\Http\Request;
-use App\Models\Author;
 
 class AuthorController extends Controller
 {
@@ -14,10 +13,11 @@ class AuthorController extends Controller
      */
     public function index()
     {
+        $author=Author::all();
         return response()->json([
-            "message" => "Daftar Author",
-            "data" => Author::all()
-        ]);
+            'status' => 200,
+            'data' => $author
+        ], 200);
     }
 
     /**
@@ -39,14 +39,18 @@ class AuthorController extends Controller
     public function store(Request $request)
     {
         $author = new Author();
-        $author->name = $request->name;
-        $author->date_of_birth = $request->date_of_birth;
-        $author->place_of_birth = $request->place_of_birth;
-        $author->gender = $request->gender;
-        $author->email = $request->email;
-        $author->hp = $request->hp;
+        $author->name = $request->input('name');
+        $author->date_of_birth = $request->input('date_of_birth');
+        $author->place_of_birth = $request->input('place_of_birth');
+        $author->gender = $request->input('gender');
+        $author->email = $request->input('email');
+        $author->no_hp = $request->input('no_hp');
         $author->save();
-        return response()->json($author, 202); 
+
+        return response()->json([
+            'status' => 201,
+            'data' => $author
+        ], 201);
     }
 
     /**
@@ -58,7 +62,18 @@ class AuthorController extends Controller
     public function show($id)
     {
         $author = Author::find($id);
-        return response()->json($author);
+
+        if($author){
+            return response()->json([
+                'status' => 200,
+                'data' => $author
+            ], 200);
+        }else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'id ' . $id . ' tidak ditemukan.'
+            ], 404);
+        }
     }
 
     /**
@@ -82,14 +97,25 @@ class AuthorController extends Controller
     public function update(Request $request, $id)
     {
         $author = Author::find($id);
-        $author->name = $request->name;
-        $author->date_of_birth = $request->date_of_birth;
-        $author->place_of_birth = $request->place_of_birth;
-        $author->gender = $request->gender;
-        $author->email = $request->email;
-        $author->hp = $request->hp;
-        $author->save();
-        return response()->json($author, 202);
+        if($author){
+            $author->name = $request->name ? $request->name : $author->name;
+            $author->date_of_birth = $request->date_of_birth ? $request->date_of_birth : $author->date_of_birth;
+            $author->place_of_birth = $request->place_of_birth ? $request->place_of_birth : $author->place_of_birth;
+            $author->gender = $request->gender ? $request->gender: $author->gender;
+            $author->email = $request->email ? $request->email : $author->email;
+            $author->no_hp = $request->no_hp ? $request->no_hp : $author->no_hp;
+            $author->save();
+
+            return response()->json([
+                'status' => 200,
+                'data' => $author
+            ], 200);
+        } else{
+            return response()->json([
+                'status' => 400,
+                'message' => 'id ' . $id . ' tidak ditemukan, gagal update data'
+            ], 404);
+        }
     }
 
     /**
@@ -100,11 +126,21 @@ class AuthorController extends Controller
      */
     public function destroy($id)
     {
-        $author = Author::findOrFail($id);
-        $author->delete();
-        return response()->json([
-            "message" => "Data Author Telah dihapus",
-            "data" => $author
-        ]);
+        $author = Author::where("id", $id)->first();
+        if($author){
+            $author->delete();
+            return response()->json([
+                'status' => 200,
+                'data' => $author
+            ], 200);
+        } else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'id ' . $id . ' tidak ditemukan, gagal menghapus data'
+            ], 404);
+        }
+    }
+    public function search($name){
+        return Author::where('name', 'like', '%'.$name.'%')->get();
     }
 }
